@@ -13,14 +13,17 @@
     ],
     "require": {
         "php": ">=5.3.2",
-        "symfony/framework-bundle": "~3.0",
-        "symfony/finder": "~3.0",
-        "symfony/yaml": "~3.0",
-        "symfony/browser-kit": "~3.0",
-        "symfony/security-bundle": "~3.0",
-        "symfony/monolog-bundle": "~2.8|~3.0",
-        "symfony/validator": "~3.0",
-        "sensio/framework-extra-bundle": "~3.0"
+        "symfony/framework-bundle": "~4.0",
+        "symfony/finder": "~4.0",
+        "symfony/yaml": "~4.0",
+        "symfony/browser-kit": "~4.0",
+        "symfony/security-bundle": "~4.0",
+        "symfony/monolog-bundle": "~2.8|~3.0|~4.0",
+        "symfony/validator": "~4.0",
+        "sensio/framework-extra-bundle": "~4.0"
+    },
+    "require-dev": {
+        "symfony/phpunit-bridge": "~4.0"
     },
     "autoload": {
         "psr-0": { "Coral\\CoreBundle": "" }
